# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Network School OAuth Provider",
    "description": "OAuth 2.0 / OpenID Connect identity provider for Network School.\n\nLets third-party apps request NS user data via standard OAuth scopes \u2014 the same pattern as \"Sign in with Google.\"\n\n## Flows\n\n- **Authorization Code + PKCE** \u2014 user-facing \"Sign in with Network School\" flow\n- **Client Credentials** \u2014 machine-to-machine, no user context\n\n## Scopes\n\n| Scope | Claims |\n|-------|--------|\n| `openid` | `sub` |\n| `email` | `email`, `email_verified` |\n| `profile` | `name`, `picture` |\n| `roles` | `roles` (Discord roles, live) |\n| `date_joined` | `date_joined` |\n| `offline_access` | refresh tokens |\n\n## Authentication\n\n- **User auth**: Discord OAuth2 (NS guild membership required)\n- **Sessions**: HS256 JWT in httponly cookie (`ns_session`)\n- **Access tokens**: RS256 JWT (auto-generated RSA keys)\n- **Client auth**: `client_secret` verified via bcrypt\n",
    "license": {
      "name": "Private",
      "url": "https://ns.com/"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/oauth/authorize": {
      "get": {
        "tags": [
          "oauth"
        ],
        "summary": "Start authorization code flow",
        "description": "Validates the client request and redirects the user to the login page (if unauthenticated) or consent page (if authenticated). The frontend renders login/consent UI, then posts back to `/oauth/authorize/consent`.",
        "operationId": "authorize_oauth_authorize_get",
        "parameters": [
          {
            "name": "response_type",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Response Type"
            }
          },
          {
            "name": "client_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Client Id"
            }
          },
          {
            "name": "redirect_uri",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Redirect Uri"
            }
          },
          {
            "name": "scope",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "openid",
              "title": "Scope"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "name": "code_challenge",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Code Challenge"
            }
          },
          {
            "name": "code_challenge_method",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Code Challenge Method"
            }
          },
          {
            "name": "nonce",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Nonce"
            }
          },
          {
            "name": "prompt",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Prompt"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "302": {
            "description": "Redirect to frontend `/login` or `/consent` with all authorize params."
          },
          "400": {
            "description": "Invalid `response_type`, unknown `client_id`, or unregistered `redirect_uri`."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/oauth/authorize/info": {
      "get": {
        "tags": [
          "oauth"
        ],
        "summary": "Get app info for consent screen",
        "description": "Returns the app name, icon, description, privacy policy URL, and requested scope details. Used by the consent page to render what permissions the user is granting.",
        "operationId": "authorize_info_oauth_authorize_info_get",
        "parameters": [
          {
            "name": "client_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Client Id"
            }
          },
          {
            "name": "scope",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "openid",
              "title": "Scope"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "App info with scope details.",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Unknown `client_id`."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/oauth/authorize/consent": {
      "post": {
        "tags": [
          "oauth"
        ],
        "summary": "Submit consent decision",
        "description": "Called by the consent page after the user clicks Allow or Deny. Returns JSON `{\"redirect_to\": url}` (not a 302) because cross-origin fetch makes the Location header inaccessible. If approved, generates an authorization code and appends it to the redirect URI. If denied, appends `error=access_denied`.",
        "operationId": "authorize_consent_oauth_authorize_consent_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_authorize_consent_oauth_authorize_consent_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "JSON with `redirect_to` URL containing either `code` + `state` (approved) or `error=access_denied` (denied).",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Unknown `client_id` or unregistered `redirect_uri`."
          },
          "401": {
            "description": "User not authenticated (no valid session cookie)."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/oauth/token": {
      "post": {
        "tags": [
          "oauth"
        ],
        "summary": "Exchange credentials for tokens",
        "description": "Token endpoint supporting three grant types:\n\n- **`client_credentials`** \u2014 Machine-to-machine. Requires `client_id` + `client_secret`. Returns an access token with no user context.\n- **`authorization_code`** \u2014 User-facing. Exchanges an authorization code (from `/oauth/authorize/consent`) for an access token. Supports PKCE (`code_verifier`). If `openid` scope was granted, also returns an `id_token`. If `offline_access` scope was granted, also returns a `refresh_token`.\n- **`refresh_token`** \u2014 Exchange a refresh token for a new access token + refresh token (rotation). Requires `client_id` + `client_secret`.\n\nClient authentication: provide `client_secret` for confidential clients, or omit for public clients (PKCE-only).",
        "operationId": "token_oauth_token_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_token_oauth_token_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Token response with `access_token`, `token_type`, `expires_in`, `scope`, and optionally `id_token`.",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Missing required fields, invalid/expired authorization code, PKCE mismatch, or unsupported grant type."
          },
          "401": {
            "description": "Invalid client credentials."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/oauth/userinfo": {
      "get": {
        "tags": [
          "oauth"
        ],
        "summary": "Get user claims (OIDC UserInfo)",
        "description": "Returns user claims based on the scopes granted to the access token.\n\nRequires a valid Bearer token in the `Authorization` header. The token must have been issued via the `authorization_code` grant (tokens from `client_credentials` have no user context and will be rejected).\n\n**Claims by scope:** `openid` \u2192 sub | `email` \u2192 email, email_verified | `profile` \u2192 name, picture | `roles` \u2192 roles | `date_joined` \u2192 date_joined.",
        "operationId": "userinfo_oauth_userinfo_get",
        "responses": {
          "200": {
            "description": "User claims object (scope-gated).",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Missing Bearer token, invalid/expired token, token has no user context, or user not found."
          }
        }
      }
    },
    "/oauth/token/introspect": {
      "post": {
        "tags": [
          "oauth"
        ],
        "summary": "Introspect a token",
        "description": "Check whether an access token is active. Returns token metadata including `scope`, `client_id`, `user_id`, `exp`, `iat`, `jti`, and `iss`. Returns `{\"active\": false}` for invalid, expired, or revoked tokens.",
        "operationId": "introspect_oauth_token_introspect_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_introspect_oauth_token_introspect_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Introspection result with `active` boolean and token metadata.",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/oauth/token/revoke": {
      "post": {
        "tags": [
          "oauth"
        ],
        "summary": "Revoke a token",
        "description": "Revokes an access token so it can no longer be used. Always returns 200 regardless of whether the token existed.",
        "operationId": "revoke_oauth_token_revoke_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_revoke_oauth_token_revoke_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Empty JSON `{}`. Token is revoked (or was already invalid).",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_authorize_consent_oauth_authorize_consent_post": {
        "properties": {
          "client_id": {
            "type": "string",
            "title": "Client Id"
          },
          "redirect_uri": {
            "type": "string",
            "title": "Redirect Uri"
          },
          "scope": {
            "type": "string",
            "title": "Scope",
            "default": "openid"
          },
          "state": {
            "type": "string",
            "title": "State",
            "default": ""
          },
          "code_challenge": {
            "type": "string",
            "title": "Code Challenge",
            "default": ""
          },
          "code_challenge_method": {
            "type": "string",
            "title": "Code Challenge Method",
            "default": ""
          },
          "nonce": {
            "type": "string",
            "title": "Nonce",
            "default": ""
          },
          "approved": {
            "type": "boolean",
            "title": "Approved"
          }
        },
        "type": "object",
        "required": [
          "client_id",
          "redirect_uri",
          "approved"
        ],
        "title": "Body_authorize_consent_oauth_authorize_consent_post"
      },
      "Body_introspect_oauth_token_introspect_post": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          },
          "client_id": {
            "type": "string",
            "title": "Client Id"
          },
          "client_secret": {
            "type": "string",
            "title": "Client Secret"
          }
        },
        "type": "object",
        "required": [
          "token"
        ],
        "title": "Body_introspect_oauth_token_introspect_post"
      },
      "Body_revoke_oauth_token_revoke_post": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          },
          "client_id": {
            "type": "string",
            "title": "Client Id"
          },
          "client_secret": {
            "type": "string",
            "title": "Client Secret"
          }
        },
        "type": "object",
        "required": [
          "token"
        ],
        "title": "Body_revoke_oauth_token_revoke_post"
      },
      "Body_token_oauth_token_post": {
        "properties": {
          "grant_type": {
            "type": "string",
            "title": "Grant Type"
          },
          "client_id": {
            "type": "string",
            "title": "Client Id"
          },
          "client_secret": {
            "type": "string",
            "title": "Client Secret"
          },
          "scope": {
            "type": "string",
            "title": "Scope"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "redirect_uri": {
            "type": "string",
            "title": "Redirect Uri"
          },
          "code_verifier": {
            "type": "string",
            "title": "Code Verifier"
          },
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          }
        },
        "type": "object",
        "required": [
          "grant_type"
        ],
        "title": "Body_token_oauth_token_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    }
  },
...
}